app/requirements.txt
toml
streamlit
streamlit_extras
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-storage
google-cloud-texttospeech
pandas
watchdog
bigframes